Finish gpu tune cleanup fixes
