+++ crates/mesh-llm-config/src/validate.rs
@@ -141,11 +141,17 @@ pub fn validate_config_diagnostics(config: &MeshConfig) -> Vec<ConfigDiagnostic>
         .defaults
         .as_ref()
         .and_then(|defaults| defaults.hardware.as_ref());
-    if let Some(defaults) = &config.defaults
-        && let Err(diagnostic) =
+    if let Some(defaults) = &config.defaults {
+        push_speculative_alias_diagnostic(
+            defaults.speculative.as_ref(),
+            "defaults.speculative",
+            &mut diagnostics,
+        );
+        if let Err(diagnostic) =
             validate_model_defaults(defaults, "defaults", config.gpu.assignment)
-    {
-        diagnostics.push(diagnostic);
+        {
+            diagnostics.push(diagnostic);
+        }
     }
     for (index, model) in config.models.iter().enumerate() {
         if model.model.trim().is_empty() {
@@ -154,6 +160,11 @@ pub fn validate_config_diagnostics(config: &MeshConfig) -> Vec<ConfigDiagnostic>
                 format!("models[{index}].model must not be empty"),
             ));
         }
+        push_speculative_alias_diagnostic(
+            model.speculative.as_ref(),
+            &format!("models[{index}].speculative"),
+            &mut diagnostics,
+        );
         if let Err(diagnostic) = validate_model_entry(
             model,
             &format!("models[{index}]"),
@@ -245,10 +256,13 @@ pub fn built_in_support_diagnostic(
 
 pub fn validate_config(config: &MeshConfig) -> Result<()> {
     let diagnostics = validate_config_diagnostics(config);
-    if diagnostics.is_empty() {
-        Ok(())
-    } else {
+    let has_errors = diagnostics
+        .iter()
+        .any(|diagnostic| diagnostic.severity == ConfigDiagnosticSeverity::Error);
+    if has_errors {
         Err(anyhow::anyhow!(legacy_validation_error_text(&diagnostics)))
+    } else {
+        Ok(())
     }
 }
 
@@ -809,19 +823,31 @@ fn validate_skippy(config: &SkippyConfig, base_path: &str) -> DiagnosticResult {
     Ok(())
 }
 
-fn validate_speculative(config: &SpeculativeConfig, base_path: &str) -> DiagnosticResult {
-    // native-mtp-n1 is a legacy alias for mtp; route it through an alias
-    // warning instead of a hard rejection so existing configs do not break.
-    if config.strategy.as_deref() == Some("native-mtp-n1") {
-        let used_path = parsed_config_path(&format!("{base_path}.strategy"))
-            .unwrap_or_else(|| ConfigPath::from_fields(["speculative", "strategy"]));
-        let canonical_path = parsed_config_path(&format!("{base_path}.strategy"))
-            .unwrap_or_else(|| ConfigPath::from_fields(["speculative", "strategy"]));
-        return Err(alias_diagnostic(used_path, canonical_path, "native-mtp-n1"));
+fn push_speculative_alias_diagnostic(
+    config: Option<&SpeculativeConfig>,
+    base_path: &str,
+    diagnostics: &mut Vec<ConfigDiagnostic>,
+) {
+    if config.and_then(|config| config.strategy.as_deref()) != Some("native-mtp-n1") {
+        return;
     }
+    let used_path = parsed_config_path(&format!("{base_path}.strategy"))
+        .unwrap_or_else(|| ConfigPath::from_fields(["speculative", "strategy"]));
+    let canonical_path = used_path.clone();
+    diagnostics.push(
+        alias_diagnostic(
+            used_path,
+            canonical_path,
+            "legacy speculative strategy native-mtp-n1 is treated as mtp",
+        )
+        .with_help("Use speculative.strategy = \"mtp\" for new config writes."),
+    );
+}
+
+fn validate_speculative(config: &SpeculativeConfig, base_path: &str) -> DiagnosticResult {
     validate_optional_enum(
         config.strategy.as_deref(),
-        &["auto", "disabled", "mtp"],
+        &["auto", "disabled", "mtp", "native-mtp-n1"],
         &format!("{base_path}.strategy"),
     )?;
     validate_optional_enum(
@@ -1796,6 +1822,34 @@ strategy = "mystery-oracle"
             diagnostics[0]
                 .message
                 .contains("defaults.speculative.strategy must be one of")
+        );
+    }
+
+    #[test]
+    fn speculative_strategy_native_mtp_n1_is_legacy_alias_warning() {
+        let config: MeshConfig = toml::from_str(
+            r#"
+[defaults.speculative]
+strategy = "native-mtp-n1"
+"#,
+        )
+        .expect("config should parse before validation");
+
+        validate_config(&config).expect("legacy speculative strategy alias should not fail");
+        let diagnostics = validate_config_diagnostics(&config);
+        assert_eq!(diagnostics.len(), 1);
+        assert_eq!(diagnostics[0].code, ConfigDiagnosticCode::AliasApplied);
+        assert_eq!(diagnostics[0].severity, ConfigDiagnosticSeverity::Warning);
+        assert_eq!(
+            diagnostics[0].path.as_ref().map(ConfigPath::render),
+            Some("defaults.speculative.strategy".to_string())
+        );
+        assert_eq!(
+            diagnostics[0]
+                .canonical_path
+                .as_ref()
+                .map(ConfigPath::render),
+            Some("defaults.speculative.strategy".to_string())
         );
     }
